add wallet api decorator
